api/pyproject.toml
[project]
name = "bookworm-api"
version = "0.1.0"
description = "API for NashTech Bookworm Assignment"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "uvicorn>=0.29.0",
    "sqlmodel>=0.0.16",
    "psycopg[binary]>=3.1.18",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "alembic>=1.13.1",
    "pydantic-settings>=2.2.1",
    "pydantic[email]>=2.11.3",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "pytest>=8.3.5",
]
